Updated navbar and search bar
Corrected the issue in the navbar where the pervious hero link in the navbar was changing width with screen resize.
Replaced it's width property with min-width: max-content.

Also, made the search input take full width of the input container, and made the search icon an absolute positioned element. Now it's consistently resizing with screen sizes.